requirements-viz.txt
streamlit>=1.28.0
snowflake-connector-python>=3.4.0
pandas>=1.5.0
plotly>=5.15.0
yfinance>=0.2.0